We refactor the env.go and add BuildRoutes:
We continue to pull the code out of main.go so it is more manageable and
easier to reason about as we build up the application.
We move the route setup and configuration to a BuildRoutes func in
application.go.

We make a slight refactor to make buildConfiguration() a private go func
that is called from Load.

And then we clean up main.go