Building out graceful shutdown:
Adding a background go routine that checks for os.signals and processes
them appropriately allowing the application to shutdown in a graceful
manner. It also provides a hook to do any type of cleanup you may want
to do in your app.

In addition, we continue to refactor the http server out of main.